refactor: move React memo logic into reusable hook
