+++ src/components/CountdownCircleTimer.jsx
@@ -1,31 +1,16 @@
 import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 import { useElapsedTime } from 'use-elapsed-time'
+import { useMemoizedProps } from '../hooks'
 import {
-  uuid,
   linearEase,
   getWrapperStyle,
   getTimeStyle,
   svgStyle,
-  getPathProps,
-  getNormalizedColors,
   getStroke,
   colorsValidator,
   visuallyHidden,
 } from '../utils'
-
-const getGradientId = (isLinearGradient, gradientUniqueKey) =>
-  isLinearGradient
-    ? `countdown-circle-timer-gradient-${gradientUniqueKey || uuid()}`
-    : ''
-
-const getStartAt = (initialRemainingTime, startAtSeconds, durationSeconds) => {
-  if (typeof initialRemainingTime === 'number') {
-    return (durationSeconds - initialRemainingTime) * 1000
-  }
-
-  return startAtSeconds ? startAtSeconds * 1000 : 0
-}
 
 const CountdownCircleTimer = (props) => {
   const {
@@ -46,27 +31,23 @@ const CountdownCircleTimer = (props) => {
     startAt: startAtSeconds,
   } = props
 
-  // useElapsedTime will make this component rerender on every frame.
-  // We memo all props that need to be computed to avoid doing that on every render
-  const { path, pathLength } = useMemo(() => getPathProps(size, strokeWidth), [
+  const {
+    path,
+    pathLength,
+    durationMilliseconds,
+    startAt,
+    normalizedColors,
+    gradientId,
+  } = useMemoizedProps({
     size,
     strokeWidth,
-  ])
-  const durationMilliseconds = useMemo(() => durationSeconds * 1000, [
     durationSeconds,
-  ])
-  const startAt = useMemo(
-    () => getStartAt(initialRemainingTime, startAtSeconds, durationSeconds),
-    [initialRemainingTime, startAtSeconds, durationSeconds]
-  )
-  const normalizedColors = useMemo(
-    () => getNormalizedColors(colors, durationMilliseconds, isLinearGradient),
-    [colors, durationMilliseconds, isLinearGradient]
-  )
-  const gradientId = useMemo(
-    () => getGradientId(isLinearGradient, gradientUniqueKey),
-    [isLinearGradient, gradientUniqueKey]
-  )
+    initialRemainingTime,
+    startAtSeconds,
+    colors,
+    isLinearGradient,
+    gradientUniqueKey,
+  })
 
   const elapsedTime = useElapsedTime(isPlaying, {
     durationMilliseconds,
