+++ src/utils/getGradientId.js
@@ -1,0 +1,6 @@
+import { uuid } from '.'
+
+export const getGradientId = (isLinearGradient, gradientUniqueKey) =>
+  isLinearGradient
+    ? `countdown-circle-timer-gradient-${gradientUniqueKey || uuid()}`
+    : ''
